Rename bigchain.py command module to bigchaindb.py
