+++ bigchaindb/commands/bigchaindb.py
